chore(packages)!: remove Java 11
Java 17 is the latest LTS release of Java and is the preferred
environment for developing Java applications. We should no longer
install Java 11 nor should we set it as the default. Note that this
change _will not_ remove Java 11 from existing installs; however, it
will change the default version to Java 17. For new installs, only Java
17 will be installed.