chore(workflows): add CI job to lint/test the playbook (#11)
This adds `ansible-lint` to the CI workflow for this repository.
Additionally, we run a single pass of the full playbook in a separate
parallel job to make sure that it runs successfully in a (hopefully)
somewhat real Ubuntu environment.

Most changes to existing files are to comply with the `ansible-lint`
rules (and the `spring` directory is just deleted entirely). We ignore
warnings that we shouldn't use the latest versions of packages or the
latest revision of a git repo -- for the most part that is actually what
we _do_ want; and adding that ignore inline everywhere would be tedious.

For the full playbook run, we start with the minimal [Ubuntu 22.04
docker container](https://hub.docker.com/_/ubuntu) and then add in some
[packages from the base WSL
image](https://cloud-images.ubuntu.com/focal/current/focal-server-cloudimg-amd64-wsl.rootfs.manifest).
Then we run the full playbook just to make sure that it succeeds. Long
term we probably want to try to run it multiple times and make sure that
`changed` is `0` at the end, but that is a future improvement.

The README is also slightly updated to highlight what `BECOME password`
means.

---------

Co-authored-by: Zachary Tucker <zachptucker@gmail.com>